deploy_flask/requirements.txt
absl-py==2.0.0
albumentations==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.0
fonttools==4.44.0
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.0.0
greenlet==3.0.1
grpcio==1.59.3
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.3
minio==7.2.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
paho-mqtt==1.6.1
pandas==2.0.3
Pillow==10.0.1
protobuf==4.25.1
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.208
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
